backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langgraph==0.2.50
langgraph-checkpoint-postgres==2.0.5
langchain-openai==0.2.14
langchain-core>=0.3.27,<0.4.0
langchain-community>=0.3.7,<0.4.0
psycopg[binary,pool]==3.2.3
tavily-python==0.5.0
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
pandas==2.2.3
openpyxl==3.1.5
python-docx==1.1.2
python-pptx==1.0.2
aiofiles==24.1.0
